requirements.txt
ultralytics
customtkinter
numpy
opencv-python
pillow
yt-dlp
